Download VCF subset from 1000 Genomes

Query the 1000 Genomes Project for a subset of their individual-level VCF files.

LD_1KG_download_vcf(
  dat,
  LD_reference = "1KGphase1",
  remote_LD = TRUE,
  vcf_folder = NULL,
  locus_dir,
  locus = NULL,
  whole_vcf = FALSE,
  download_method = "axel",
  force_new_vcf = FALSE,
  query_by_regions = FALSE,
  remove_tmps = TRUE,
  nThread = 1,
  conda_env = "echoR",
  verbose = TRUE
)

Arguments

dat
GWAS summary statistics subset to query the LD panel with.

LD_reference
LD reference to use:

"1KGphase1" : 1000 Genomes Project Phase 1

"1KGphase3" : 1000 Genomes Project Phase 3

"UKB" : Pre-computed LD from a British European-decent subset of UK Biobank.

remote_LD
Whether to pull the LD reference from remote repository, or locally stored files.

locus_dir
Storage directory to use.

download_method

"axel" : Multi-threaded

"wget" : Single-threaded

"download.file" : Single-threaded

"internal" : Single-threaded (passed to download.file)

"wininet" : Single-threaded (passed to download.file)

"libcurl" : Single-threaded (passed to download.file)

"curl" : Single-threaded (passed to download.file)

or "download.file" (single-threaded) .

query_by_regions
You can make queries with tabix in two different ways:

query_by_regions=F (default)

Return a vcf with all positions between the min/max in dat. Takes up more storage but is MUCH faster

query_by_regions=T

Return a vcf with only the exact positions present in dat. Takes up less storage but is MUCH slower

remove_tmps
Remove all temporary files like VCF, npz, and plink files.

nThread
Number of threads to parallelize over.

conda_env
Conda environment name.

verbose
Print messages.

Details

data("BST1") vcf_subset.popDat <- LD_1KG_download_vcf( dat = BST1, LD_reference = "1KGphase1", locus_dir = file.path(tempdir(), locus_dir) )

See also