Procure an LD matrix for fine-mapping

Calculate and/or query linkage disequilibrium (LD) from reference panels (UK Biobank, 1000 Genomes), or user-supplied datasets.

load_or_create(
  locus_dir,
  dat,
  force_new_LD = FALSE,
  LD_reference = "1KGphase1",
  LD_genome_build = "hg19",
  superpopulation = "EUR",
  remote_LD = TRUE,
  download_method = "direct",
  vcf_folder = NULL,
  LD_block = FALSE,
  LD_block_size = 0.7,
  remove_correlates = FALSE,
  fillNA = 0,
  verbose = TRUE,
  server = FALSE,
  remove_tmps = TRUE,
  as_sparse = TRUE,
  conda_env = "echoR",
  nThread = 1
)

Arguments

dat
The locus subset of the full summary stats file.

force_new_LD
By default, if an LD matrix file for a given locus is already present, then echolocatoR will just use the preexisting file. Set force_new_LD=T to override this and extract a new subset.

LD_reference
Which linkage disequilibrium reference panel do you want to use. Options include:

"UKB"

A pre-caclulated LD reference matrix from a subset of caucasian British individuals from the UK Biobank. See Wiessbrod et al. (2019) for more details.

"1KGphase1"

Download a subset of the 1000 Genomes Project Phase 1 vcf and calculate LD on the fly with plink.

"1KGphase3"

Download a subset of the 1000 Genomes Project Phase 3 vcf and calculate LD on the fly with plink.

"<path>/*.vcf" or "<path>/*.vcf.gz"

Alternatively, users can provide their own custom panel by supplying a list of .vcf file path (one per locus) which echolocatoR will use to compute LD (using plink).

superpopulation
Subset your LD reference panel by superopulation. Setting the superpopulation is not currently possible when LD_reference="UKB". 1KGphase1 options include:

"AFR"

African [descent]

"AMR"

Ad-mixed American

"EAS"

East Asian

"EUR"

European

"SAS"

South Asian

remote_LD
When acquiring LD matrices, the default is to delete the full vcf or npz files after echolocatoR has extracted the necssary subset. However, if you wish to keep these full files (which can be quite large) set remote_LD=T.

LD_block
Calculate LD blocks with plink and only include the block to which the lead SNP belongs.

LD_block_size
Adjust the granularity of block sizes when LD_block=T.

remove_correlates
A named list, where the names are the RSIDs of SNPs whose LD correlates you wish to remove, and the value is the absolute r2 threshold you wish to filter at for each RSID respectively (e.g. remove_correlates = c("rs76904798"=.2, "rs10000737"=.8)). This will also remove the SNPs in remove_correlates themselves.

verbose
Whether echolocatoR should be verbose or silent.

server
Whether echolocatoR is being run on a computing cluster/server or on a local machine.

remove_tmps
Whether to remove any temporary files (e.g. FINEMAP output files) after the pipeline is done running.

conda_env
The name of a conda environment to use.

Value

A symmetric LD matrix of pairwise r values.

Details

Options:

Download pre-computed LD matrix from UK Biobank.

Download raw vcf file from 1KG and compute LD on the fly.

Compute LD on the fly from a user-supplied vcf file.

Use a user-supplied pre-computed LD-matrix.

See also

Examples

if (FALSE) {
data("BST1")
data("locus_dir")
locus_dir <- file.path(tempdir(), locus_dir)
BST1 <- BST1[seq(1,50),]

LD_matrix <- load_or_create(locus_dir = locus_dir, 
                            dat = BST1, 
                            LD_reference = "1KGphase1") 
}