Get LD using snpStats package

Get LD using snpStats package

snpstats_get_LD(
  LD_folder,
  plink_prefix = "plink",
  select.snps = NULL,
  stats = c("R"),
  symmetric = TRUE,
  depth = "max",
  nThread = 1,
  verbose = TRUE
)

Arguments

LD_folder
Locus-specific LD output folder.

stats
A character vector specifying the linkage disequilibrium measures to be calculated. This should contain one or more of the strings: "LLR", "OR", "Q", "Covar", "D.prime", "R.squared", ad "R"

symmetric
When no y argument is supplied this argument controls the format of the output band matrices. If TRUE, symmetric matrices are returned and, otherwise, an upper triangular matrices are returned

depth
When y is not supplied, this parameter is mandatory and controls the maximum lag between columns of x considered. Thus, LD statistics are calculated between x[,i] and x[,j] only if i and j differ by no more than depth

Source

snpStats Bioconductor page LD tutorial

See also

Examples

dat <- data.table::fread("/pd-omics/brian/Fine_Mapping/Data/GWAS/Kunkle_2019/ABCA7/Multi-finemap/ABCA7.Kunkle_2019.1KGphase3_LD.Multi-finemap.tsv.gz")

#> Error in data.table::fread("/pd-omics/brian/Fine_Mapping/Data/GWAS/Kunkle_2019/ABCA7/Multi-finemap/ABCA7.Kunkle_2019.1KGphase3_LD.Multi-finemap.tsv.gz"): File '/pd-omics/brian/Fine_Mapping/Data/GWAS/Kunkle_2019/ABCA7/Multi-finemap/ABCA7.Kunkle_2019.1KGphase3_LD.Multi-finemap.tsv.gz' does not exist or is non-readable. getwd()=='/Users/schilder/Desktop/echoverse/echoLD/docs/reference'

LD_folder <- "/pd-omics/brian/Fine_Mapping/Data/GWAS/Kunkle_2019/ABCA7/LD"
LD_matrix <- snpstats_get_LD(LD_folder = LD_folder, select.snps = dat$SNP)

#> Error in snpstats_get_LD(LD_folder = LD_folder, select.snps = dat$SNP): could not find function "snpstats_get_LD"